Remove TS exercises
