+++ .gitignore
@@ -78,7 +78,4 @@ typings/
 .env.development.local
 .env.test.local
 .env.production.local
-
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
+.idea
