--- src/react-excercises/CounterExercise.js
@@ -1,6 +1,0 @@
-import React, { useState } from "react";
-import "../App.css";
-
-export default function ClickCounterExercise() {
-  return <p>CounterExercise</p>;
-}
