--- src/react-excercises/HumanfriendlyDateCountdownExercise.js
@@ -1,6 +1,0 @@
-import React, { useState, useEffect } from "react";
-import "../App.css";
-
-export default function HumanfriendlyDateCountdownExercise() {
-	return <p>HumanfriendlyDateCountdownExercise</p>;
-}
