--- src/react-excercises/ReducerCounterExercise.js
@@ -1,6 +1,0 @@
-import React, { useState, useReducer } from "react";
-import "../App.css";
-
-export default function ClickCounterExercise() {
-	return <p>ReducerCounterExercise</p>;
-}
