+++ src/react-exercises/Counter.css
@@ -1,0 +1,36 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 2rem;
+}
+
+.counter {
+  font-size: 8rem;
+  margin-top: -1rem;
+}
+
+.button {
+  border: none;
+  border-radius: 6px;
+  background-color: #eab8b2;
+  color: #231f20;
+  padding: 0.5em 1em;
+  margin: 4px;
+  min-width: 44px;
+  min-height: 44px;
+  font-weight: 700;
+}
+
+.decrease-button {
+  background-color: #dd5928;
+}
+
+.increase-button {
+  background-color: #00b3b0;
+}
+
+.reset-button {
+  background-color: #e3d9d7;
+  min-width: 96px;
+}
