+++ src/react-exercises/DateCountdownExercise.js
@@ -1,5 +1,5 @@
-import React, { useState, useEffect } from "react";
-import "../App.css";
+import React from "react";
+import "./Counter.css";
 
 export default function DateCountdownExercise() {
   return <p>DateCountdownExercise</p>;
