+++ src/react-exercises/HumanfriendlyDateCountdownExercise.js
@@ -1,0 +1,6 @@
+import React from "react";
+import "./Counter.css";
+
+export default function HumanfriendlyDateCountdownExercise() {
+  return <p>HumanfriendlyDateCountdownExercise</p>;
+}
