--- src/ts-excercises/excercise-4.tsx
@@ -1,69 +1,0 @@
-import React from "react";
-
-/*
-
-Intro:
-
-    As we introduced "type" to both User and Admin
-    it's now easier to distinguish between them.
-    Once object type checking logic was extracted
-    into separate functions isUser and isAdmin -
-    logPerson function got new type errors.
-
-Exercise:
-
-    Figure out how to help TypeScript understand types in
-    this situation and apply necessary fixes.
-
-*/
-
-interface User {
-  type: 'user';
-  name: string;
-  age: number;
-  occupation: string;
-}
-
-interface Admin {
-  type: 'admin';
-  name: string;
-  age: number;
-  role: string;
-}
-
-export type Person = User | Admin;
-
-export const persons: Person[] = [
-  { type: 'user', name: 'Max Mustermann', age: 25, occupation: 'Chimney sweep' },
-  { type: 'admin', name: 'Jane Doe', age: 32, role: 'Administrator' },
-  { type: 'user', name: 'Kate Müller', age: 23, occupation: 'Astronaut' },
-  { type: 'admin', name: 'Bruce Willis', age: 64, role: 'World saver' }
-];
-
-export function isAdmin(person: Person) {
-  return person.type === 'admin';
-}
-
-export function isUser(person: Person) {
-  return person.type === 'user';
-}
-
-// TODO: remove the comments in the code
-export default function Excercise4() {
-  return <div>
-    <h4>TS Excercise 4</h4>
-    {/* {persons.map((person) => {
-      let additionalInformation: string;
-      if (isAdmin(person)) {
-        additionalInformation = person.role;
-      }
-      if (isUser(person)) {
-          additionalInformation = person.occupation;
-      }
-      return <div>{`Name: ${person.name}, Age: ${person.age}, Additional Information: ${additionalInformation}`}</div>
-    })} */}
-  </div>
-}
-
-// In case if you are stuck:
-// https://www.typescriptlang.org/docs/handbook/advanced-types.html#using-type-predicates
