--- src/ts-excercises/excercise-5.tsx
@@ -1,100 +1,0 @@
-import React from "react";
-
-/*
-
-Intro:
-
-    Time to filter the data! In order to be flexible
-    we filter users using a number of criteria and
-    return only those matching all of the criteria.
-    We don't need Admins yet, we only filter Users.
-
-Exercise:
-
-    Without duplicating type structures, modify
-    filterUsers function definition so that we can
-    pass only those criteria which are needed,
-    and not the whole User information as it is
-    required now according to typing.
-
-Higher difficulty bonus exercise:
-
-    Exclude "type" from filter criterias.
-
-*/
-
-interface User {
-  type: 'user';
-  name: string;
-  age: number;
-  occupation: string;
-}
-
-interface Admin {
-  type: 'admin';
-  name: string;
-  age: number;
-  role: string;
-}
-
-export type Person = User | Admin;
-
-export const persons: Person[] = [
-  { type: 'user', name: 'Max Mustermann', age: 25, occupation: 'Chimney sweep' },
-  {
-      type: 'admin',
-      name: 'Jane Doe',
-      age: 32,
-      role: 'Administrator'
-  },
-  {
-      type: 'user',
-      name: 'Kate Müller',
-      age: 23,
-      occupation: 'Astronaut'
-  },
-  {
-      type: 'admin',
-      name: 'Bruce Willis',
-      age: 64,
-      role: 'World saver'
-  },
-  {
-      type: 'user',
-      name: 'Wilson',
-      age: 23,
-      occupation: 'Ball'
-  },
-  {
-      type: 'admin',
-      name: 'Agent Smith',
-      age: 23,
-      role: 'Administrator'
-  }
-];
-
-export const isAdmin = (person: Person): person is Admin => person.type === 'admin';
-export const isUser = (person: Person): person is User => person.type === 'user';
-
-export function filterUsers(persons: Person[], criteria: User): User[] {
-  return persons.filter(isUser).filter((user) => {
-      const criteriaKeys = Object.keys(criteria) as (keyof User)[];
-      return criteriaKeys.every((fieldName) => {
-          return user[fieldName] === criteria[fieldName];
-      });
-  });
-}
-
-// TODO: remove the comments in the code
-export default function Excercise5() {
-  return <div>
-    <h4>TS Excercise 5</h4>
-    {/* {filterUsers( persons, {age: 23}).map((person) => {
-      return <div>{`Name: ${person.name}, Age: ${person.age}`}</div>
-    })} */}
-  </div>
-}
-
-// In case if you are stuck:
-// https://www.typescriptlang.org/docs/handbook/advanced-types.html#mapped-types
-// https://www.typescriptlang.org/docs/handbook/release-notes/typescript-2-8.html#predefined-conditional-types
